pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chift"
version = "0.1.0"
description = "Odoo Integration API - Sync contacts and invoices from Odoo to local database"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "apscheduler>=3.10.4",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",  # Pin to 4.x - bcrypt 5.0 has breaking changes with passlib
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.21",
]

[tool.setuptools.packages.find]
include = ["app*", "scripts*"]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
    "pytest>=8.0.0",
    "ty>=0.0.12",
]

[tool.ruff]
# Target Python 3.13+
target-version = "py313"

# Line length
line-length = 100

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "build",
    "dist",
    "*.egg-info",
    "alembic/versions",  # Exclude auto-generated migration files
    ".idea",
    ".vscode",
]

[tool.ruff.lint]
# Enable rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "TRY",    # tryceratops
    "FLY",    # flynt
    "PERF",   # perflint
    "RUF",    # ruff-specific rules
]

# Disable specific rules that may be too strict
ignore = [
    "ISC001",   # Conflicts with ruff formatter
    "COM812",   # Conflicts with ruff formatter
    "E501",     # Line too long (handled by formatter)
    "TRY003",   # Long messages in exception constructors
    "EM101",    # Exception string literals
    "EM102",    # Exception f-string literals
    "TRY301",   # Abstract raise to inner function
    "TRY300",   # Consider moving statement to else block
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value comparison
    "PT009",    # Use unittest assertions (prefer pytest style)
    "T201",     # Print found (allow print statements)
    "G004",     # Logging f-string (prefer f-strings for readability)
    "G201",     # Logging .exception should be used instead of .error with exc_info
    "B008",     # Function call in argument defaults (FastAPI Depends pattern)
    "B904",     # raise from err (allow raise without from)
    "ARG001",   # Unused function argument (common in FastAPI/SQLAlchemy)
    "ARG002",   # Unused method argument
    "TRY400",   # Use logging.exception (allow logger.error)
    "RUF012",   # Mutable class attributes (SQLAlchemy models)
    "TC001",    # Move application imports to type-checking block
    "PLC0415",  # Import should be at top level (allow dynamic imports)
    "TRY002",   # Create custom exception (allow generic Exception)
    "SIM102",   # Use single if statement (readability preference)
]

[tool.ruff.lint.per-file-ignores]
# Allow certain patterns in specific files
"alembic/env.py" = ["ERA001", "T201", "PTH120", "I001"]  # Allow patterns in alembic config
"app/core/config.py" = ["UP007"]  # Allow Optional[] syntax for settings
"scripts/*.py" = ["T201"]  # Allow prints in utility scripts

[tool.ruff.lint.isort]
known-first-party = ["app", "scripts"]
force-sort-within-sections = true

[tool.ruff.lint.pylint]
max-args = 7
max-branches = 15

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Use Unix line endings
line-ending = "auto"
